refactor : xml 읽기 / 쓰기 및 수정 기능 분리
